Added GlobalExceptionHandler class and response classes for Hotel, Room and Reservation

+++ src/main/java/com/example/hotelbookingapp/model/Reservation.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -22,7 +24,7 @@ public class Reservation {
 
     private LocalDate checkInDate;
     private LocalDate checkOutDate;
-    private double totalPrice;
+    private BigDecimal totalPrice;
 
     @ManyToOne
     @JoinColumn(name = "room_id")
